Invert object/type_traits dependency layering
